requirements.txt
# ── Core API ──────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
scalar-fastapi>=0.1.0
python-multipart>=0.0.9

# ── Vector database ───────────────────────────────────────────
qdrant-client>=1.9.0

# ── Scheduler ────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Audio & feature extraction ────────────────────────────────
librosa>=0.11.0
soundfile>=0.13.0
soxr>=1.0.0
pydub>=0.25.0
audioread>=3.0.0
scipy>=1.11.0
numpy>=1.26.0
numba>=0.59.0

# ── Deep learning inference ───────────────────────────────────
onnxruntime>=1.18.0
vibenet>=0.1.0

# ── YouTube download ─────────────────────────────────────────
yt-dlp>=2025.1.1

# ── HTTP / utils ─────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0
